Fix block generation by updating codec versions in wasm (#810)
* Fixes for staging testnet

* Fix codec versioning

* Comment unneeded trace

* Update lib.rs

* Update parity-codec

* Add script to update wasm deps and update them.

* Update hashes

* Add BBQ Birch testnet

* it's default-features, apparently.

* Use simpler version spec for parity codec

* Typo

+++ node/primitives/src/lib.rs
@@ -54,7 +54,7 @@ pub type AccountId = ::primitives::H256;
 pub type AccountIndex = u32;
 
 /// Balance of an account.
-pub type Balance = u64;
+pub type Balance = u128;
 
 /// The Ed25519 pub key of an session that belongs to an authority of the chain. This is
 /// exactly equivalent to what the substrate calls an "authority".
